ocr-service-python/requirements.txt
Flask==3.0.2
easyocr==1.7.2
pillow==11.0.0
torch==2.5.1+cpu
